Enable parser.c on Windows
I was originally planning to use mmap on posix systems which clearly doesn't work on windows hence the guards but I settled on just reading it directly which Windows does support.
I would expect this to fail if building with msvc because Microsoft prepended _ to all supported posix functions but mingw-w64 supports them. I think there is a define that brings them back on msvc, otherwise a bunch of macros can be used to rename everything back.
The gdi frontend still crashes on start but the command line one is working provided Windows Terminal is used.