Initial work to remove hardcoded custom screen creation
Anything related to what happens on a screen should be in
the json file except anything needing to be done in code.
Currently the mapping between screens and screen specific
setup code is still in code but moving it from the switch
statement to an array will make moving the mapping to the
json file easier in future.