Partially fix support for mobaxterm
Previously mingw-w64 builds would appear to hang on start up, turns out this was
because of stdout not flushing on print. After fixing that, user input was being
ignored because of some weird issue with getch so I switched to setting up stdin
the same way as on unix likes and then using read. This worked fine with msys2's
version of mintty but then failed under mobaxterm with GetConsoleMode not liking
the stdin handle. Until this can be fixed, I have added a fallback where setting
up stdin's mode can be skipped for the mingw-w64 build if GetConsoleMode errors.
As a result, using mobaxterm currently requires pressing enter after each input.

Currently, the plan is to remove the gdi frontend in favour of an old version of
mobaxterm which still has support for windows xp and maybe even 2000 which would
allow just using the cmd version. I also tried other terminals which had support
for xp like conemu but was unable to correctly display unicode chars outside the
bmp even with windows 10. The gdi frontend was added as I thought having support
for older versions of windows would be cool and because it would allow me to get
some experience using gdi. However, keeping feature parity with the cmd frontend
using a modern terminal would take too much time away from actual game dev.