Merge shared files into backend
Orignally, the split was used to indicate which files were
part of the backend and which were also used by the tools.
However, the only tool that was using any shared files was
printgamedata.c which is basically a cutdown frontend so I
don't feel like the distinction was worth keeping anymore.