Switch from c17 to c99 to support building for old windows versions
The older versions of cygwin I am planning to have gcc 3.4.4 so no
support for new c features.

EDIT: This commit was moved to a different branch in case it ends
up being useful. I had forgotten that the Ubuntu 22.04 version of
mingw-w64 gcc makes binaries that run fine on Windows 2000. There
are some issues with text printing on NT 4.0 SP4 but these should
be fixable.