Add initially equipped equipment to json
Treat equipment array in json not being 63 items as an error for both
the validator and at runtime.
Modify SetEquippedItems in equipment.c to allow it to be called from
LoadDefaultPlayerInfo in parser.c which only has a struct PlayerInfo
instead of a struct GameInfo.
Add missing equipment.h dependancy for save.c.