Changing equipment during combat takes a turn
Added a new field and enum in CombatEventInfo to track the action
performed including the existing attack action and a new equipment
swap action which triggers on successful swapping(i.e. not if there
was only one item in a category).

Added a function UpdateCombat to add an event to the log for equipment
swapping and moved existing code for checking if the player has won to
it which required two new fields: changedEquipment and playerWon to
make these accessable elsewhere.

Fixed some existing error detection and reporting in CreateCombatString.