Fix building on Windows
Both Comso and Binutil copies of AR cause gzip to be called in a way
that fails on Windows betcause of some issues with std handles if
called by dash so rename ar to ar.com and call directly.

Cleaning zstd on Windows via it's clean target has always been broken
so just remove the lib/obj directory instead.