Stop capturing errors in extensions
The catch was meant to only handle issues with getModule
not being able to find the file, not any issues the file
may have.

Arrays directly defined in the Extension field rather
than in a seperate field are now supports.
It turns out that just setting generated array's type
to `null` and relying on Haxe to fill it in works fine.