Add bounds checking to password loading
To get an upper bound for equipment, I needed Type.getEnumConstructs
which only works for actual Enums, not abstract ones that act closer
to C enums with each enum constructor mapping to an integer.
For saving, we still need to be able to convert to and from ints but
Type.enumIndex and Type.createEnumIndex allow doing this.